compounds compiles after refactor to spectral properties (finally 😫)

+++ spike/models/compound/phase/gas/CompletedGas.hpp
@@ -17,7 +17,7 @@ namespace phase {
         relation::GasPropertyStateRelation<si::thermal_conductivity> thermal_conductivity;
         relation::GasPropertyStateRelation<si::dynamic_viscosity> dynamic_viscosity;
         field::CompletedStateField<si::density> density;
-        relation::PolynomialRailyardRelation<si::wavenumber,float,0,1> refractive_index;
+        relation::PolynomialRailyardRelation<si::wavenumber,double,0,1> refractive_index;
     };
 
     /*
