+++ spike/models/compound/phase/liquid/PartlyKnownLiquid.hpp
@@ -18,8 +18,8 @@ namespace phase {
         field::OptionalStateField<si::density> density;
         field::OptionalStateField<si::pressure> vapor_pressure;
         field::OptionalStateField<si::surface_energy> surface_tension;
-        relation::OptionalStateRelation<relation::PolynomialRailyardRelation<si::wavenumber,float,0,1>> refractive_index;
-        relation::OptionalStateRelation<relation::PolynomialRailyardRelation<si::wavenumber,float,0,1>> extinction_coefficient;
+        relation::OptionalStateRelation<relation::PolynomialRailyardRelation<si::wavenumber,double,0,1>> refractive_index;
+        relation::OptionalStateRelation<relation::PolynomialRailyardRelation<si::wavenumber,double,0,1>> extinction_coefficient;
 
         /*
         Return a `PartlyKnownLiquid` that has the properties of `known` where present, otherwise substitute with properties of `base`
