+++ spike/models/compound/phase/solid/CompletedSolid.hpp
@@ -17,8 +17,8 @@ namespace phase {
         field::CompletedStateField<si::dynamic_viscosity> dynamic_viscosity;
         field::CompletedStateField<si::density> density;
         field::CompletedStateField<si::pressure> vapor_pressure;
-        relation::PolynomialRailyardRelation<si::wavenumber,float,0,1> refractive_index;
-        relation::PolynomialRailyardRelation<si::wavenumber,float,0,1> extinction_coefficient;
+        relation::PolynomialRailyardRelation<si::wavenumber,double,0,1> refractive_index;
+        relation::PolynomialRailyardRelation<si::wavenumber,double,0,1> extinction_coefficient;
         relation::PolynomialRailyardRelation<si::wavenumber,si::attenuation,0,1> absorption_coefficient;
 
         field::CompletedStateField<si::pressure> bulk_modulus;     // K   bulk modulus
